requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
deprecation==2.1.0
fastapi==0.102.0
frozenlist==1.4.0
gotrue==1.0.4
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
langchain==0.0.273
langsmith==0.0.26
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
packaging==23.1
postgrest==0.10.8
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
realtime==1.0.0
requests==2.31.0
six==1.16.0
slack-sdk==3.21.3
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tenacity==8.2.3
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
websockets==10.4
yarl==1.9.2
